+++ services/http/HttpServer.js
@@ -1,6 +1,6 @@
 import http from 'node:http';
-import path from 'node:path';
-import * as url from 'node:url';
+// import path from 'node:path';
+// import * as url from 'node:url';
 import express from 'express';
 import RequestLoggerMiddleware from './middleware/RequestLogger.js';
 import I18nMiddleware from './middleware/I18n.js';
@@ -18,12 +18,12 @@ class HttpServer extends Base {
     super(app);
     this.express = express();
     this.express.disable('x-powered-by');
-    const dirname = url.fileURLToPath(new URL('.', import.meta.url));
-    this.express.set('views', [
-      this.app.foldersConfig.views,
-      path.join(dirname, '../../views'),
-    ]);
-    this.express.set('view engine', 'pug');
+    // const dirname = url.fileURLToPath(new URL('.', import.meta.url));
+    // this.express.set('views', [
+    //   this.app.foldersConfig.views,
+    //   path.join(dirname, '../../views'),
+    // ]);
+    // this.express.set('view engine', 'pug');
 
     this.express.use(new PrepareAppInfoMiddleware(this.app).getMiddleware());
     this.express.use(new IpDetector(this.app).getMiddleware());