+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 5.0.0-alpha.13
+
+[UPDATE] update deps
+[UPDATE] update i18n internal implementation
+[CHANGE] disable https server view
+
 ### 5.0.0-alpha.12
 
 [UPDATE] update deps